Remove the NationalGeographics class
There are only 3 functions which are the core of this program. I revised the
code today to find it's not necessary to split them to two classess. Also, I
added the xdm member to the PhotoOfTheDay class, which indicates the X window
manager of the environment. Perhaps we can port it to KDE seemlessly now.